Simple change to Row.Iterator and Row.Extractor which should allow for a fairly intuitive use of both text and binary formats.
Binary formats utilize the serializer format, except in the case of
nextString() which always returns a string. next() always returns the
typed value. Binary formats are a "programmer centric" view of the
database.

Text formats always used the string format, except in the case of
primitive extractors (nextInt(), nextDouble() etc.). next() always
returns the string value. Text is the "human centric" view of the
database.

Yes, this means that programmers are not humans.